requirements.txt
oracledb
pyjson5
python-dotenv
rich
